requirements.txt
loguru~=0.7.2
aiohttp~=3.9.1
asyncpg~=0.29.0
aiofile~=3.8.8
Faker~=22.2.0
argon2-cffi~=23.1.0
python-dotenv~=1.0.1
